Move measure return type to common
