+++ src/reanimated2/NativeMethods.ts
@@ -1,6 +1,7 @@
 /* global _WORKLET _measure _scrollTo _dispatchCommand _setGestureState */
 import { Component } from 'react';
 import { findNodeHandle } from 'react-native';
+import { MeasuredDimensions } from './commonTypes';
 import { RefObjectFunction } from './hook/commonTypes';
 import { shouldBeUseWeb } from './PlatformChecker';
 
@@ -8,15 +9,6 @@ export function getTag(
   view: null | number | React.Component<any, any> | React.ComponentClass<any>
 ): null | number {
   return findNodeHandle(view);
-}
-
-export interface MeasuredDimensions {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-  pageX: number;
-  pageY: number;
 }
 
 const isNative = !shouldBeUseWeb();
